refactor: update oracle integration and frontend components for testnet deployment
- Revised oracle deployment strategy to utilize a self-deployed OracleMock for testnet simplicity, replacing the Chainlink-compatible aggregator approach.
- Enhanced frontend components (OracleCard, TokenFaucetCard, VaultActions) to reflect changes in oracle integration and improve user experience.
- Updated configuration files to handle missing artifacts gracefully during builds.
- Improved type safety in API requests and contract interactions across various components and hooks.
- Adjusted bot configuration parameters for lenders, borrowers, and vault users to optimize performance and resource allocation.

+++ frontend/src/hooks/useMarketInsights.ts
@@ -68,7 +68,7 @@ export function useMarketInsights(): MarketInsights {
       totalBorrowShares: marketDataStruct.totalBorrowShares,
       lastUpdate: marketDataStruct.lastUpdate,
       fee: marketDataStruct.fee,
-      price: oraclePrice,
+      price: oraclePrice as bigint,
       // rateAtTarget: undefined - SDK handles basic IRM automatically
     });
